[NightShift] Homepage
Add some Homepage content that will be easy to customize.

Act as a profesional modern designer.

Don't forget to use our guidelines from docs/ui_design.

This is for a template, so nothing specific.

It should have all that modern homepage require, CTA, features, hero, things like that.

Also, don't forget the footer with the sitemap.

In all, this is for a template, make sure it has everything needed for SEO, google search engine, etc.

Don't forget to add the "pricing"/"plans" - should be a copy with the tabs.

Generated by NightShift Task #34