[NightShift] New Subscribe
In the footer, next to the site roadmap, there should be a "subscribe" input field so that users can subscribe to receive news, and things like that.
Make sure it is Turbo, so the "Subscribed! Thank you." message appears directly where the input field is.
So, there should be a new Subscribe or something like that table in our database, that has only an email for now I think?
Also, we need a new page/view, let's say /subscribe for now, so that we can put it as the root/homepage when we are building the app and can collect the email of interested people, if that makes sense. Something like "Coming Soon! Get on the waiting list". Think of some nice message to let the user know this thing is in progress and ask him to add his email. See what I mean?

Generated by NightShift Task #59